split worker logic into threads allowing to safely utilize parallel setting and removing the overhead of stream switching
